Reduce duplication in Blockchain implementation
